organiser/src/main/java/com/task/organiser/service/TaskService.java
package com.task.organiser.service;

import com.task.organiser.entity.Comment;
import com.task.organiser.entity.Task;
import com.task.organiser.repository.CommentRepository;
import com.task.organiser.repository.TaskRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

import java.time.LocalDateTime;
import java.util.List;
import java.util.Optional;

@Service
public class TaskService {
    @Autowired
    private TaskRepository taskRepository;
    @Autowired
    private CommentRepository commentRepository;

    public Task createTask(Task task) {
        return taskRepository.save(task);
    }

    public Task getTaskById(Long id) {
        return taskRepository.findById(id).orElse(null);
    }

    public List<Task> getTasksByAssigneeId(Long assigneeId) {
        return taskRepository.findByAssigneeId(assigneeId);
    }

    public Task updateTask(Long id, Task task) {
        task.setId(id);
        return taskRepository.save(task);
    }
    public Task transitionTask(Long id, String status) {
        Task task = taskRepository.findById(id).orElseThrow(() -> new RuntimeException("Task not found"));
        task.setStatus(status);
        return taskRepository.save(task);
    }

    public Task addCommentToTask(Long taskId, String commentContent) {
        Task task = taskRepository.findById(taskId)
            .orElseThrow(() -> new RuntimeException("Task not found"));

        Comment comment = new Comment();
        comment.setComment(commentContent);
        comment.setCreatedDate(LocalDateTime.now());
        comment.setTask(task);
        task.addComment(comment);
        commentRepository.save(comment);
        return taskRepository.save(task);
    }
    public void deleteTask(Long id) {
        taskRepository.deleteById(id);
    }
}
